docs: injectpaymentonion timestamps are u32
injectpaymentonion reports success via injectonion_succeed(),
which emits both timestamps with json_add_u32 from the u32
wallet_payment struct members:

  lightningd/pay.c:1769-1770
    json_add_u32(response, "created_at", payment->timestamp);
    json_add_u32(response, "completed_at", *payment->completed_at);
  wallet/wallet.h:363-364
    u32 timestamp;
    u32 *completed_at;

Changelog-None

+++ cln-rpc/src/model.rs
@@ -12938,8 +12938,8 @@ pub mod responses {
 
 	#[derive(Clone, Debug, Deserialize, Serialize)]
 	pub struct InjectpaymentonionResponse {
-	    pub completed_at: u64,
-	    pub created_at: u64,
+	    pub completed_at: u32,
+	    pub created_at: u32,
 	    pub created_index: u64,
 	    pub payment_preimage: Secret,
 	}
